Add files via upload
The nonparametric method for failures diagnosis in the aircraft control system that uses the measurements of the control signals and the aircraft states only. It doesn’t require a priori information of the aircraft model parameters, training or statistical calculations, and is based on an analytical nonparametric one-step-ahead state prediction approach. This makes it possible to predict the behaviour of unidentified and failure dynamic systems, weaken the requirements to control signals and reduce the diagnostic time and problem complexity.